backend/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-mfa2==2.5.0
django-easy-audit==1.3.5
Pillow==10.1.0
python-decouple==3.8
cryptography==39.0.0
redis==5.0.1
celery==5.3.4
requests==2.31.0
openai==1.3.7
PyJWT==2.8.0
python-magic==0.4.27
gunicorn==21.2.0
whitenoise==6.6.0